amazon_scraper/requirements.txt
pymongo
requests
beautifulsoup4
python-dotenv
pandas
lxml
urllib3
